requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
geoalchemy2==0.15.2
psycopg2-binary==2.9.9
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
weasyprint
jinja2
